Commit message:
Add comment_count to article table and adjust constraints

In the database migration scripts, a new column 'comment_count' was added to the 'article' table. Also, the constraints for the article and member tables were adjusted and a 'FOREIGN KEY' constraint was explicitly added to the 'article' table referencing the member 'id'.

Signed-off-by: seungju.kim <pipiru100@gmail.com>