Add Prometheus configuration to deployment workflow
This change introduces an additional step in the Gradle deployment workflow which involves transfer of the Prometheus configuration file. The file is sent to the temporary location on the deployment server, and then moved to its final destination (`/etc/prometheus/prometheus.yml`). This ensures proper monitoring set up when deploying new versions of the application.

Signed-off-by: seungju.kim <pipiru100@gmail.com>